Batch process messages if needed
Added an index to the messages table, this allows the ORDER BY to be
internalized to the storage, which makes queries faster. Not that it
matters terribly.

Changed the message reading loop to read all overdue messages from the
same query result set before yielding the task.

Also took the opportunity to clean up the naming some, and refactor the
select! so that it's not repeated, instead modifying the sleep duration
before requesting.